Move supervisor tests to separate module
